Skills/이메일 스크랩/requirements-gmail.txt
# Gmail → md 스크립트용 (gmail_to_md.py)
google-api-python-client>=2.100.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.1.1
